skill-flow/data/skills-refined-skillclaw-36k/skillsmp/office-to-md/requirements.txt
python-docx>=0.8.11
PyMuPDF>=1.23.0
Pillow>=10.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
python-pptx>=1.0.0  # 需要1.0.0+版本以支持Python 3.11+
flask>=3.0.0
flask-cors>=4.0.0
flask-restx>=1.3.0
